Moved logging to AzToolsFramework, working with MatEditor and SMC
