+++ Code/Framework/AzToolsFramework/AzToolsFramework/aztoolsframework_files.cmake
@@ -147,6 +147,8 @@ set(FILES
     Entity/SliceEditorEntityOwnershipServiceBus.h
     Fingerprinting/TypeFingerprinter.h
     Fingerprinting/TypeFingerprinter.cpp
+	Logger/AzQtTraceLogger.cpp
+	Logger/AzQtTraceLogger.h
     Manipulators/AngularManipulator.cpp
     Manipulators/AngularManipulator.h
     Manipulators/BaseManipulator.cpp
