+++ Gems/Atom/Tools/MaterialEditor/Code/Source/MaterialEditorApplication.cpp
@@ -352,43 +352,7 @@ namespace MaterialEditor
             ExitMainLoop();
         }
     }
-    /*
-    void MaterialEditorApplication::WriteStartupLog()
-    {
-        using namespace AzFramework;
 
-        AZ::IO::FileIOBase* fileIO = AZ::IO::FileIOBase::GetInstance();
-        AZ_Assert(fileIO != nullptr, "FileIO should be running at this point");
-
-        // There is no log system online so we have to create your own log file.
-        char resolveBuffer[AZ_MAX_PATH_LEN] = { 0 };
-        fileIO->ResolvePath("@user@", resolveBuffer, AZ_MAX_PATH_LEN);
-
-        // Note: @log@ hasn't been set at this point
-        AZStd::string logDirectory;
-        StringFunc::Path::Join(resolveBuffer, "log", logDirectory);
-        fileIO->SetAlias("@log@", logDirectory.c_str());
-
-        fileIO->CreatePath("@root@");
-        fileIO->CreatePath("@user@");
-        fileIO->CreatePath("@log@");
-
-        AZStd::string logPath;
-        StringFunc::Path::Join(logDirectory.c_str(), "MaterialEditor.log", logPath);
-
-        m_logFile.reset(aznew LogFile(logPath.c_str()));
-        if (m_logFile)
-        {
-            m_logFile->SetMachineReadable(false);
-            for (const LogMessage& message : m_startupLogSink)
-            {
-                m_logFile->AppendLog(LogFile::SEV_NORMAL, message.window.c_str(), message.message.c_str());
-            }
-            m_startupLogSink = {};
-            m_logFile->FlushLog();
-        }
-    }
-    */
     void MaterialEditorApplication::LoadSettings()
     {
         AZ::SerializeContext* context = nullptr;
@@ -476,8 +440,7 @@ namespace MaterialEditor
             return;
         }
 
-        m_traceLogger->WriteStartupLog("MaterialEditor.log");
-        //WriteStartupLog();
+        m_traceLogger.WriteStartupLog("MaterialEditor.log");
 
         if (!LaunchDiscoveryService())
         {
