+++ Gems/Atom/Tools/MaterialEditor/Code/Source/MaterialEditorApplication.h
@@ -20,9 +20,8 @@
 #include <AzFramework/Logging/LogFile.h>
 #include <AzToolsFramework/API/AssetDatabaseBus.h>
 #include <AzToolsFramework/API/EditorPythonConsoleBus.h>
-
+#include <AzToolsFramework/Logger/AzQtTraceLogger.h>
 #include <AzQtComponents/Application/AzQtApplication.h>
-#include <AzQtComponents/Application/AzQtTraceLogger.h>
 
 #include <QTimer>
 
@@ -106,7 +105,6 @@ namespace MaterialEditor
         void CompileCriticalAssets();
 
         void ProcessCommandLine(const AZ::CommandLine& commandLine);
-        //void WriteStartupLog();
 
         void LoadSettings();
         void UnloadSettings();
@@ -125,6 +123,8 @@ namespace MaterialEditor
 
         AZStd::vector<LogMessage> m_startupLogSink;
         AZStd::unique_ptr<AzFramework::LogFile> m_logFile;
+
+        AzToolsFramework::AzQtTraceLogger m_traceLogger;
 
         //! Local user settings are used to store material browser tree expansion state
         AZ::UserSettingsProvider m_localUserSettings;
