+++ Gems/Atom/Tools/ShaderManagementConsole/Code/Source/ShaderManagementConsoleApplication.cpp
@@ -39,8 +39,6 @@
 #include <AzCore/Settings/SettingsRegistryMergeUtils.h>
 #include <AzCore/Utils/Utils.h>
 
-#include <AzQtComponents/Application/AzQtTraceLogger.h>
-
 AZ_PUSH_DISABLE_WARNING(4251 4800, "-Wunknown-warning-option") // disable warnings spawned by QT
 #include <QFileInfo>
 #include <QObject>
@@ -344,6 +342,8 @@ namespace ShaderManagementConsole
         {
             return;
         }
+
+        m_traceLogger.WriteStartupLog("ShaderManagementConsole.log");
 
         //[GFX TODO][ATOM-415] Try to factor out some of this stuff with AtomSampleViewerApplication
         AzToolsFramework::AssetDatabase::AssetDatabaseRequestsBus::Handler::BusConnect();
